2023 PP1 Project1 Delete 디버깅

+++ src/main/java/com/pp1/project1/WordCRUD.java
@@ -57,7 +57,7 @@ public class WordCRUD implements ICRUD {
         System.out.print("=> 정말로 삭제하실래요?(Y/N) ");
         String ans = s.next();
         if(ans.equalsIgnoreCase("y")){
-            list.remove(idList.get(id-1));
+            list.remove((int) idList.get(id-1));
             System.out.println("단어가 삭제되었습니다. ");
         }
         else System.out.println("취소되었습니다. ");
